+++ wfrp/src/main/java/com/warhammer/wfrp/age/AgeDto.java
@@ -1,0 +1,5 @@
+package com.warhammer.wfrp.age;
+
+public record AgeDto(String age) {
+
+}
