+++ wfrp/src/main/java/com/warhammer/wfrp/age/AgeService.java
@@ -1,8 +1,5 @@
-package com.warhammer.wfrp.service;
+package com.warhammer.wfrp.age;
 
-import com.warhammer.wfrp.dto.appearance.AgeDto;
-import com.warhammer.wfrp.engine.AgeGenerator;
-import com.warhammer.wfrp.mapper.AgeMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
